feat: Add Payment Repository Interface
