+++ src/core/repositories/subscriptions/payment/index.ts
@@ -1,4 +1,4 @@
-import { PaymentRepository } from "./interface";
+import { PaymentRepository } from "./payment-repository.interface";
 import { PrismaPaymentRepository } from "./prisma";
 
 export const getPaymentRepository = (): PaymentRepository => {
